共通して使うメソッドを共通化する
