+++ app/controllers/talks/action_uncompleted_controller.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 class Talks::ActionUncompletedController < ApplicationController
+  include SearchUser
   before_action :require_admin_login
 
   def index
@@ -17,16 +18,6 @@ class Talks::ActionUncompletedController < ApplicationController
       ).page(params[:page])
     else
       @talks = @talks.page(params[:page])
-    end
-  end
-
-  private
-
-  def validate_search_word(search_word)
-    if search_word.match?(/^[\w-]+$/)
-      search_word.strip if search_word.strip.length >= 3
-    elsif search_word.strip.length >= 2
-      search_word.strip
     end
   end
 end
